backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0
apscheduler==3.10.4
tenacity==9.0.0
python-dotenv==1.0.1
orjson==3.10.10

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.7.1
